[CF] AUDIT-30 Создать сервис для работы с типами событий (audit_event_type) https://jira.i-novus.ru/browse/AUDIT-30

+++ audit-service/src/main/java/ru/i_novus/ms/audit/service/EventTypeService.java
@@ -38,7 +38,7 @@ public class EventTypeService {
             eventTypeRepository.save(EventTypeEntity
                     .builder()
                     .name(name)
-                    .auditTypeId(Integer.valueOf(auditType))
+                    .auditTypeId(Short.valueOf(auditType))
                     .build());
         }
     }
